# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Analysis": {
        "additionalProperties": false,
        "properties": {
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "title": "Matched Skills",
            "type": "array"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "title": "Missing Skills",
            "type": "array"
          },
          "provider": {
            "enum": [
              "local",
              "gemini",
              "openai"
            ],
            "title": "Provider",
            "type": "string"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/Suggestion"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "provider",
          "matched_skills",
          "missing_skills",
          "suggestions"
        ],
        "title": "Analysis",
        "type": "object"
      },
      "AnalyzeRequest": {
        "additionalProperties": false,
        "properties": {
          "jd_text": {
            "maxLength": 15000,
            "minLength": 30,
            "title": "Jd Text",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "jd_text",
          "revision"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "Body_upload_resume_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_upload_post",
        "type": "object"
      },
...
      "Education": {
        "additionalProperties": false,
        "properties": {
          "degree": {
            "default": "",
            "maxLength": 300,
            "title": "Degree",
            "type": "string"
          },
          "institution": {
            "default": "",
            "maxLength": 300,
            "title": "Institution",
            "type": "string"
          },
          "year_end": {
            "default": 0,
            "maximum": 2200.0,
            "minimum": 0.0,
            "title": "Year End",
            "type": "integer"
          },
          "year_start": {
            "default": 0,
            "maximum": 2200.0,
            "minimum": 0.0,
            "title": "Year Start",
            "type": "integer"
          }
        },
        "title": "Education",
        "type": "object"
      },
      "Experience": {
        "additionalProperties": false,
        "properties": {
          "achievements": {
            "items": {
              "type": "string"
            },
            "maxItems": 40,
            "title": "Achievements",
            "type": "array"
          },
          "company": {
            "default": "",
            "maxLength": 300,
            "title": "Company",
            "type": "string"
          },
          "date_from": {
            "default": "",
            "maxLength": 50,
            "title": "Date From",
            "type": "string"
          },
          "date_to": {
            "default": "",
            "maxLength": 50,
            "title": "Date To",
            "type": "string"
          },
          "location": {
            "default": "",
            "maxLength": 200,
            "title": "Location",
            "type": "string"
          },
          "responsibilities": {
            "default": "",
            "maxLength": 6000,
            "title": "Responsibilities",
            "type": "string"
          },
          "role": {
            "default": "",
            "maxLength": 300,
            "title": "Role",
            "type": "string"
          }
        },
        "title": "Experience",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobError": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "JobError",
        "type": "object"
      },
      "JobView": {
        "additionalProperties": true,
        "description": "`result` has the same shape as the synchronous response of the operation.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "result": {
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "operation",
          "status"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "Project": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 6000,
            "title": "Description",
            "type": "string"
          },
          "tech": {
            "items": {
              "type": "string"
            },
            "maxItems": 60,
            "title": "Tech",
            "type": "array"
          },
          "title": {
            "default": "",
            "maxLength": 300,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "Project",
        "type": "object"
      },
...
      "ResumeFields": {
        "additionalProperties": false,
        "properties": {
          "certificates": {
            "default": "",
            "maxLength": 3000,
            "title": "Certificates",
            "type": "string"
          },
          "education": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/Education"
                },
                "type": "array"
              }
            ],
            "default": "",
            "maxLength": 3000,
            "title": "Education"
          },
          "email": {
            "default": "",
            "maxLength": 200,
            "title": "Email",
            "type": "string"
          },
          "experience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/Experience"
                },
                "type": "array"
              }
            ],
            "default": "",
            "maxLength": 12000,
            "title": "Experience"
          },
          "full_name": {
            "default": "",
            "maxLength": 120,
            "title": "Full Name",
            "type": "string"
          },
          "languages": {
            "default": "",
            "maxLength": 500,
            "title": "Languages",
            "type": "string"
          },
          "location": {
            "default": "",
            "maxLength": 200,
            "title": "Location",
            "type": "string"
          },
          "phone": {
            "default": "",
            "maxLength": 100,
            "title": "Phone",
            "type": "string"
          },
          "position": {
            "default": "",
            "maxLength": 200,
            "title": "Position",
            "type": "string"
          },
          "projects": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/Project"
                },
                "type": "array"
              }
            ],
            "default": "",
            "maxLength": 6000,
            "title": "Projects"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "maxItems": 60,
            "title": "Skills",
            "type": "array"
          },
          "summary": {
            "default": "",
            "maxLength": 3000,
            "title": "Summary",
            "type": "string"
          }
        },
        "title": "ResumeFields",
        "type": "object"
      },
      "ResumeRecord": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Analysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/ResumeFields"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "jd_text": {
            "default": "",
            "title": "Jd Text",
            "type": "string"
          },
          "lifecycle": {
            "default": "draft",
            "enum": [
              "draft",
              "active",
              "archived"
            ],
            "title": "Lifecycle",
            "type": "string"
          },
          "resume_id": {
            "title": "Resume Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "enum": [
              "extracted",
              "edited",
              "reviewed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "resume_id",
          "filename",
          "fields",
          "status",
          "revision",
          "created_at"
        ],
        "title": "ResumeRecord",
        "type": "object"
      },
...
      "SaveRequest": {
        "additionalProperties": false,
        "properties": {
          "fields": {
            "$ref": "#/components/schemas/ResumeFields"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "fields",
          "revision"
        ],
        "title": "SaveRequest",
        "type": "object"
      },
...
      "Suggestion": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "maxLength": 900,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "kind": {
            "enum": [
              "strength",
              "improvement",
              "gap"
            ],
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "detail",
          "kind"
        ],
        "title": "Suggestion",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Upload, review, compare and export a factual resume. Start with /api/session.",
    "title": "CareerBot",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/resume/upload": {
      "post": {
        "operationId": "upload_resume_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload"
      }
    },
    "/resume/{resume_id}": {
      "delete": {
        "operationId": "delete_resume__resume_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete"
      },
      "get": {
        "operationId": "get_resume_resume__resume_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Resume"
      }
    },
...
    "/resume/{resume_id}/improve": {
      "post": {
        "operationId": "improve_resume__resume_id__improve_post",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Accepted as a background job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Improve"
      }
    },
...
    "/resume/{resume_id}/pdf": {
      "get": {
        "operationId": "pdf_resume__resume_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "template",
            "required": false,
            "schema": {
              "default": "modern",
              "enum": [
                "modern",
                "classic",
                "minimalist"
              ],
              "title": "Template",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pdf"
      }
    },
    "/resume/{resume_id}/save": {
      "post": {
        "operationId": "save_resume__resume_id__save_post",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save"
      }
    },
...
  },
  "servers": [
    {
      "description": "Docker gateway",
      "url": "/api"
    }
  ]
}